feat udistrital/arka_cliente#584: Ajuste Contable
- Mejora manejo de modelos
