+++ swagger/swagger.json
@@ -277,7 +277,7 @@
                     "200": {
                         "description": "",
                         "schema": {
-                            "$ref": "#/definitions/ajustesHelper.DetalleAjuste"
+                            "$ref": "#/definitions/models.DetalleAjuste"
                         }
                     },
                     "403": {
@@ -1406,97 +1406,9 @@
         }
     },
     "definitions": {
-        "ajustesHelper.DetalleAjuste": {
-            "title": "DetalleAjuste",
-            "type": "object",
-            "properties": {
-                "Movimiento": {
-                    "$ref": "#/definitions/models.Movimiento"
-                },
-                "TrContable": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/ajustesHelper.DetalleMovimientoContable"
-                    }
-                }
-            }
-        },
-        "ajustesHelper.DetalleCuenta": {
-            "title": "DetalleCuenta",
-            "type": "object",
-            "properties": {
-                "Codigo": {
-                    "type": "string"
-                },
-                "Nombre": {
-                    "type": "string"
-                },
-                "RequiereTercero": {
-                    "type": "boolean"
-                }
-            }
-        },
-        "ajustesHelper.DetalleMovimientoContable": {
-            "title": "DetalleMovimientoContable",
-            "type": "object",
-            "properties": {
-                "Credito": {
-                    "type": "number",
-                    "format": "double"
-                },
-                "Cuenta": {
-                    "$ref": "#/definitions/ajustesHelper.DetalleCuenta"
-                },
-                "Debito": {
-                    "type": "number",
-                    "format": "double"
-                },
-                "Descripcion": {
-                    "type": "string"
-                },
-                "TerceroId": {
-                    "$ref": "#/definitions/models.Tercero"
-                }
-            }
-        },
-        "ajustesHelper.PreMovAjuste": {
-            "title": "PreMovAjuste",
-            "type": "object",
-            "properties": {
-                "Credito": {
-                    "type": "number",
-                    "format": "double"
-                },
-                "Cuenta": {
-                    "type": "string"
-                },
-                "Debito": {
-                    "type": "number",
-                    "format": "double"
-                },
-                "Descripcion": {
-                    "type": "string"
-                },
-                "TerceroId": {
-                    "type": "integer",
-                    "format": "int64"
-                }
-            }
-        },
         "ajustesHelper.PreTrAjuste": {
             "title": "PreTrAjuste",
-            "type": "object",
-            "properties": {
-                "Descripcion": {
-                    "type": "string"
-                },
-                "Movimientos": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/ajustesHelper.PreMovAjuste"
-                    }
-                }
-            }
+            "type": "object"
         },
         "map[string]interface{}": {
             "title": "map[string]interface{}",
@@ -1798,6 +1710,21 @@
                 }
             }
         },
+        "models.DetalleAjuste": {
+            "title": "DetalleAjuste",
+            "type": "object",
+            "properties": {
+                "Movimiento": {
+                    "$ref": "#/definitions/models.Movimiento"
+                },
+                "TrContable": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/models.DetalleMovimientoContable"
+                    }
+                }
+            }
+        },
         "models.DetalleBaja": {
             "title": "DetalleBaja",
             "type": "object",
@@ -1830,6 +1757,21 @@
                 },
                 "TipoBaja": {
                     "type": "string"
+                }
+            }
+        },
+        "models.DetalleCuenta": {
+            "title": "DetalleCuenta",
+            "type": "object",
+            "properties": {
+                "Codigo": {
+                    "type": "string"
+                },
+                "Nombre": {
+                    "type": "string"
+                },
+                "RequiereTercero": {
+                    "type": "boolean"
                 }
             }
         },
@@ -1918,6 +1860,29 @@
                     "items": {
                         "$ref": "#/definitions/models.DetalleTercero"
                     }
+                }
+            }
+        },
+        "models.DetalleMovimientoContable": {
+            "title": "DetalleMovimientoContable",
+            "type": "object",
+            "properties": {
+                "Credito": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "Cuenta": {
+                    "$ref": "#/definitions/models.DetalleCuenta"
+                },
+                "Debito": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "Descripcion": {
+                    "type": "string"
+                },
+                "TerceroId": {
+                    "$ref": "#/definitions/models.Tercero"
                 }
             }
         },
